+++ lib/pdf/core.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require_relative 'core/io_wrapper'
 require_relative 'core/pdf_object'
 require_relative 'core/annotations'
 require_relative 'core/byte_string'